move lib folder inside src
