+++ src/pages/api/create-topic.ts
@@ -1,6 +1,6 @@
 import { NextApiRequest, NextApiResponse } from "next";
 import { Topic } from "@prisma/client";
-import prisma from "lib/prisma.server";
+import prisma from "@lib/prisma.server";
 
 export default async function handler(
   req: NextApiRequest,
