+++ src/pages/issues/[slug].tsx
@@ -6,10 +6,10 @@ import Layout from "@components/layout/layout";
 import Link from "@components/shared/link";
 import MasonryLayout from "@components/shared/masonry/masonry-layout";
 import { GetStaticPaths, GetStaticProps, NextPage } from "next";
-import prisma from "lib/prisma.server";
-import { getIssue } from "lib/getters/unique-getters.server";
-import { slugsToPaths } from "lib/helpers.server";
-import { getIssues } from "lib/getters/many-getters.server";
+import prisma from "@lib/prisma.server";
+import { getIssue } from "@lib/getters/unique-getters.server";
+import { slugsToPaths } from "@lib/helpers.server";
+import { getIssues } from "@lib/getters/many-getters.server";
 import NothingHereWrapper from "@components/Latest/nothing-here-wrapper";
 
 const Issue: NextPage<IssuePageType> = ({
@@ -36,7 +36,8 @@ const Issue: NextPage<IssuePageType> = ({
           border="1px solid black"
           borderRadius="0.5rem"
           _hover={{ background: "rgba(222, 222, 222, 0.8)" }}
-          _active={{ background: "transparent" }}>
+          _active={{ background: "transparent" }}
+        >
           Read the PDF!!
         </Link>
       )}
