+++ tsconfig.json
@@ -20,6 +20,7 @@
       "@components/*": ["./src/components/*"],
       "@theme/*": ["./src/theme/*"],
       "@server/*": ["./src/server/*"],
+      "@lib/*": ["./src/lib/*"],
       "@/*": ["./src/*"]
     }
   },
